# backend/swagger/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/settings": {
            "get": {
                "description": "Returns the current configuration settings for signup, user directories, rules, frontend.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Settings"
                ],
                "summary": "Get system settings",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Property to retrieve: `userDefaults`, `frontend`, `auth`, `server`, `sources`",
                        "name": "property",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "System settings data",
                        "schema": {
                            "$ref": "#/definitions/settings.Settings"
                        }
                    }
                }
            }
        },
        "/api/settings/config": {
            "get": {
                "description": "Returns the current running configuration settings as YAML format with optional comments and filtering.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "Settings"
                ],
                "summary": "Get system settings as YAML",
                "parameters": [
                    {
                        "type": "boolean",
                        "description": "Show all values (true) or only non-default values (false, default)",
                        "name": "full",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Include comments in YAML (true) or plain YAML (false, default)",
                        "name": "comments",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "System settings in YAML format",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "settings.Auth": {
            "type": "object",
            "properties": {
                "adminPassword": {
                    "description": "secret: the password of the admin user. If not set, the default is \"admin\".",
                    "type": "string"
                },
                "adminUsername": {
                    "description": "secret: the username of the admin user. If not set, the default is \"admin\".",
                    "type": "string"
                },
                "key": {
                    "description": "secret: the key used to sign the JWT tokens. If not set, a random key will be generated.",
                    "type": "string"
                },
                "methods": {
                    "$ref": "#/definitions/settings.LoginMethods"
                },
                "tokenExpirationHours": {
                    "description": "time in hours each web UI session token is valid for. Default is 2 hours.",
                    "type": "integer"
                },
                "totpSecret": {
                    "description": "secret: secret used to encrypt TOTP secrets",
                    "type": "string"
                }
            }
        },
...
        "settings.CustomTheme": {
            "type": "object",
            "properties": {
                "css": {
                    "description": "The css file path and filename to use for the theme.",
                    "type": "string"
                },
                "description": {
                    "description": "The description of the theme to display in the UI.",
                    "type": "string"
                }
            }
        },
        "settings.Database": {
            "type": "object",
            "properties": {
                "migrateFrom": {
                    "description": "path to old BoltDB database file for migration (optional)",
                    "type": "string"
                },
                "path": {
                    "description": "path to SQLite database file",
                    "type": "string"
                }
            }
        },
        "settings.ExternalLink": {
            "type": "object",
            "required": [
                "text",
                "url"
            ],
            "properties": {
                "text": {
                    "description": "the text to display on the link",
                    "type": "string"
                },
                "title": {
                    "description": "the title to display on hover",
                    "type": "string"
                },
                "url": {
                    "description": "the url to link to",
                    "type": "string"
                }
            }
        },
...
        "settings.Filesystem": {
            "type": "object",
            "required": [
                "createDirectoryPermission",
                "createFilePermission"
            ],
            "properties": {
                "createDirectoryPermission": {
                    "description": "Unix permissions like 755, 2755, 1777 (default: 755)",
                    "type": "string"
                },
                "createFilePermission": {
                    "description": "Unix permissions like 644, 755, 2755 (default: 644)",
                    "type": "string"
                }
            }
        },
        "settings.Frontend": {
            "type": "object",
            "properties": {
                "description": {
                    "description": "description that shows up in html head meta description",
                    "type": "string"
                },
                "disableDefaultLinks": {
                    "description": "disable default links in the sidebar",
                    "type": "boolean"
                },
                "disableNavButtons": {
                    "description": "disable the nav buttons in the sidebar",
                    "type": "boolean"
                },
                "disableUsedPercentage": {
                    "description": "disable used percentage for the sources in the sidebar",
                    "type": "boolean"
                },
                "externalLinks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/settings.ExternalLink"
                    }
                },
                "favicon": {
                    "description": "path to a favicon to use for the frontend",
                    "type": "string"
                },
                "loginButtonText": {
                    "description": "text to display on the login button",
                    "type": "string"
                },
                "loginIcon": {
                    "description": "path to an image file for the login page icon",
                    "type": "string"
                },
                "name": {
                    "description": "display name",
                    "type": "string"
                },
                "oidcLoginButtonText": {
                    "description": "text to display on the OIDC login button",
                    "type": "string"
                },
                "styling": {
                    "$ref": "#/definitions/settings.StylingConfig"
                }
            }
        },
        "settings.IndexSqlConfig": {
            "type": "object",
            "properties": {
                "batchSize": {
                    "description": "number of items to batch in a single transaction, typically 500-5000. higher = faster but could use more memory.",
                    "type": "integer"
                },
                "cacheSizeMB": {
                    "description": "size of the SQLite cache in MB",
                    "type": "integer"
                },
                "disableReuse": {
                    "description": "enable to always create a new indexing database on startup.",
                    "type": "boolean"
                },
                "startupIntegrityCheck": {
                    "description": "the method used to check the integrity of the index database on startup (default: quickCheck)",
                    "type": "string",
                    "enum": [
                        "quickCheck",
                        "probe",
                        "off"
                    ]
                },
                "walMode": {
                    "description": "enable the more complex WAL journaling mode. Slower, more memory usage, but better for deployments with constant user activity.",
                    "type": "boolean"
                }
            }
        },
        "settings.Integrations": {
            "type": "object",
            "properties": {
                "media": {
                    "$ref": "#/definitions/settings.Media"
                },
                "office": {
                    "$ref": "#/definitions/settings.OnlyOffice"
                }
            }
        },
...
        "settings.LogConfig": {
            "type": "object",
            "properties": {
                "apiLevels": {
                    "description": "separated list of log levels to enable for the API. (eg. \"info|warning|error\")",
                    "type": "string"
                },
                "json": {
                    "description": "output in json format",
                    "type": "boolean"
                },
                "levels": {
                    "description": "separated list of log levels to enable. (eg. \"info|warning|error|debug\")",
                    "type": "string"
                },
                "noColors": {
                    "description": "disable colors in the output",
                    "type": "boolean"
                },
                "output": {
                    "description": "output location. (eg. \"stdout\" or \"path/to/file.log\")",
                    "type": "string"
                },
                "utc": {
                    "description": "use UTC time in the output instead of local time",
                    "type": "boolean"
                }
            }
        },
        "settings.LoginMethods": {
            "type": "object",
            "properties": {
                "jwt": {
                    "$ref": "#/definitions/settings.JwtAuthConfig"
                },
                "ldap": {
                    "$ref": "#/definitions/settings.LdapConfig"
                },
                "noauth": {
                    "description": "if set to true, overrides all other auth methods and disables authentication",
                    "type": "boolean"
                },
                "oidc": {
                    "$ref": "#/definitions/settings.OidcConfig"
                },
                "password": {
                    "$ref": "#/definitions/settings.PasswordAuthConfig"
                },
                "proxy": {
                    "$ref": "#/definitions/settings.ProxyAuthConfig"
                }
            }
        },
        "settings.Media": {
            "type": "object",
            "properties": {
                "convert": {
                    "description": "config for ffmpeg conversion settings",
                    "allOf": [
                        {
                            "$ref": "#/definitions/settings.FfmpegConvert"
                        }
                    ]
                },
                "debug": {
                    "description": "output ffmpeg stdout for media integration -- careful can produces lots of output!",
                    "type": "boolean"
                },
                "exiftoolPath": {
                    "description": "path to exiftool executable",
                    "type": "string"
                },
                "extractEmbeddedSubtitles": {
                    "description": "extract embedded subtitles from media files",
                    "type": "boolean"
                },
                "ffmpegPath": {
                    "description": "path to ffmpeg directory with ffmpeg and ffprobe (eg. /usr/local/bin)",
                    "type": "string"
                }
            }
        },
...
        "settings.OnlyOffice": {
            "type": "object",
            "required": [
                "secret",
                "url"
            ],
            "properties": {
                "internalUrl": {
                    "description": "An optional internal address that the filebrowser server can use to communicate with the OnlyOffice Document Server, could be useful to bypass proxy.",
                    "type": "string"
                },
                "secret": {
                    "description": "secret: authentication key for OnlyOffice integration",
                    "type": "string"
                },
                "url": {
                    "description": "The URL to the OnlyOffice Document Server, needs to be accessible to the user.",
                    "type": "string"
                },
                "viewOnly": {
                    "description": "view only mode for OnlyOffice",
                    "type": "boolean"
                }
            }
        },
...
        "settings.Server": {
            "type": "object",
            "required": [
                "sources"
            ],
            "properties": {
                "baseURL": {
                    "description": "base URL for the server, the subpath that the server is running on.",
                    "type": "string"
                },
                "cacheDir": {
                    "description": "path to the cache directory, used for thumbnails and other cached files",
                    "type": "string"
                },
                "cacheDirCleanup": {
                    "description": "whether to automatically cleanup the cache directory. Note: docker must also mount a persistent volume to persist the cache (default: false)",
                    "type": "boolean"
                },
                "database": {
                    "description": "SQLite database configuration",
                    "allOf": [
                        {
                            "$ref": "#/definitions/settings.Database"
                        }
                    ]
                },
                "disablePreviewResize": {
                    "description": "disable resizing of previews for faster loading over slow connections",
                    "type": "boolean"
                },
                "disablePreviews": {
                    "description": "disable all previews thumbnails, simple icons will be used",
                    "type": "boolean"
                },
                "disableTypeDetectionByHeader": {
                    "description": "disable type detection by header, useful if filesystem is slow.",
                    "type": "boolean"
                },
                "disableUpdateCheck": {
                    "description": "disables backend update check service",
                    "type": "boolean"
                },
                "disableWebDAV": {
                    "description": "disable webdav support (default: false)",
                    "type": "boolean"
                },
                "externalUrl": {
                    "description": "used by share links if set (eg. http://mydomain.com)",
                    "type": "string"
                },
                "filesystem": {
                    "description": "filesystem settings",
                    "allOf": [
                        {
                            "$ref": "#/definitions/settings.Filesystem"
                        }
                    ]
                },
                "indexSqlConfig": {
                    "description": "Index database SQL configuration",
                    "allOf": [
                        {
                            "$ref": "#/definitions/settings.IndexSqlConfig"
                        }
                    ]
                },
                "internalUrl": {
                    "description": "used by integrations if set, this is the base domain that an integration service will use to communicate with filebrowser (eg. http://localhost:8080)",
                    "type": "string"
                },
                "listen": {
                    "description": "address to listen on (default: 0.0.0.0)",
                    "type": "string"
                },
                "logging": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/settings.LogConfig"
                    }
                },
                "maxArchiveSize": {
                    "description": "maximum archive/unarchive size in GB. 0 means no limit. (default: 20)",
                    "type": "integer"
                },
                "minSearchLength": {
                    "description": "minimum length of search query to begin searching (default: 3)",
                    "type": "integer"
                },
                "numImageProcessors": {
                    "description": "number of concurrent image processing jobs used to create previews, default is number of cpu cores available.",
                    "type": "integer"
                },
                "port": {
                    "description": "port to listen on",
                    "type": "integer"
                },
                "socket": {
                    "description": "socket to listen on",
                    "type": "string"
                },
                "sources": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/settings.Source"
                    }
                },
                "tlsCert": {
                    "description": "path to TLS cert",
                    "type": "string"
                },
                "tlsKey": {
                    "description": "path to TLS key",
                    "type": "string"
                }
            }
        },
        "settings.Settings": {
            "type": "object",
            "properties": {
                "auth": {
                    "$ref": "#/definitions/settings.Auth"
                },
                "frontend": {
                    "$ref": "#/definitions/settings.Frontend"
                },
                "integrations": {
                    "$ref": "#/definitions/settings.Integrations"
                },
                "server": {
                    "$ref": "#/definitions/settings.Server"
                },
                "userDefaults": {
                    "$ref": "#/definitions/settings.UserDefaults"
                }
            }
        },
        "settings.Source": {
            "type": "object",
            "required": [
                "path"
            ],
            "properties": {
                "config": {
                    "$ref": "#/definitions/settings.SourceConfig"
                },
                "name": {
                    "description": "display name",
                    "type": "string"
                },
                "path": {
                    "description": "file system path. (Can be relative)",
                    "type": "string"
                }
            }
        },
...
        "settings.StylingConfig": {
            "type": "object",
            "properties": {
                "customCSS": {
                    "description": "if a valid path to a css file is provided, it will be applied for all users. (eg. \"reduce-rounded-corners.css\")",
                    "type": "string"
                },
                "customThemes": {
                    "description": "A list of custom css files that each user can select to override the default styling. if \"default\" is key name then it will be the default option.",
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/settings.CustomTheme"
                    }
                },
                "darkBackground": {
                    "description": "Specify a valid CSS color property value to use as the background color in dark mode",
                    "type": "string"
                },
                "disableEventThemes": {
                    "description": "disable the event based themes,",
                    "type": "boolean"
                },
                "lightBackground": {
                    "description": "specify a valid CSS color property value to use as the background color in light mode",
                    "type": "string"
                }
            }
        },
        "settings.UserDefaults": {
            "type": "object",
            "properties": {
                "customTheme": {
                    "description": "Name of theme to use chosen from custom themes config.",
                    "type": "string"
                },
                "darkMode": {
                    "description": "should dark mode be enabled",
                    "type": "boolean"
                },
                "dateFormat": {
                    "description": "when false, the date is relative, when true, the date is an exact timestamp",
                    "type": "boolean"
                },
                "debugOffice": {
                    "description": "debug onlyoffice editor",
                    "type": "boolean"
                },
                "deleteAfterArchive": {
                    "description": "delete source files after successful creation/extraction of archives",
                    "type": "boolean"
                },
                "deleteWithoutConfirming": {
                    "description": "delete files without confirmation",
                    "type": "boolean"
                },
                "disableOfficePreviewExt": {
                    "description": "deprecated: use disablePreviewExt instead",
                    "type": "string"
                },
                "disableOnlyOfficeExt": {
                    "description": "list of file extensions to disable onlyoffice editor for",
                    "type": "string"
                },
                "disablePreviewExt": {
                    "description": "space separated list of file extensions to disable preview for",
                    "type": "string"
                },
                "disableQuickToggles": {
                    "description": "disable the quick toggles in the sidebar",
                    "type": "boolean"
                },
                "disableSearchOptions": {
                    "description": "disable the search options in the search bar",
                    "type": "boolean"
                },
                "disableSettings": {
                    "description": "disable the user from viewing the settings page",
                    "type": "boolean"
                },
                "disableUpdateNotifications": {
                    "description": "disable update notifications banner for admin users",
                    "type": "boolean"
                },
                "disableViewingExt": {
                    "description": "space separated list of file extensions to disable viewing for",
                    "type": "string"
                },
                "editorQuickSave": {
                    "description": "show quick save button in editor",
                    "type": "boolean"
                },
                "fileLoading": {
                    "description": "upload and download settings",
                    "allOf": [
                        {
                            "$ref": "#/definitions/users.FileLoading"
                        }
                    ]
                },
                "gallerySize": {
                    "description": "0-9 - the size of the gallery thumbnails",
                    "type": "integer"
                },
                "hideFilesInTree": {
                    "description": "hide files in the sidebar tree navigation, when true, will show only directories.",
                    "type": "boolean"
                },
                "hideSidebarFileActions": {
                    "description": "hide the file actions in the sidebar",
                    "type": "boolean"
                },
                "locale": {
                    "description": "language to use: eg. de, en, or fr",
                    "type": "string"
                },
                "lockPassword": {
                    "description": "disable the user from changing their password",
                    "type": "boolean"
                },
                "loginMethod": {
                    "description": "login method to use: eg. password, proxy, oidc",
                    "type": "string"
                },
                "permissions": {
                    "$ref": "#/definitions/settings.UserDefaultsPermissions"
                },
                "preferEditorForMarkdown": {
                    "description": "prefer editor first for markdown files instead of the Markdown Viewer.",
                    "type": "boolean"
                },
                "preview": {
                    "$ref": "#/definitions/settings.UserDefaultsPreview"
                },
                "quickDownload": {
                    "description": "show icon to download in one click",
                    "type": "boolean"
                },
                "showHidden": {
                    "description": "show hidden files in the UI. On windows this includes files starting with a dot and windows hidden files",
                    "type": "boolean"
                },
                "showSelectMultiple": {
                    "description": "show select multiple files on desktop",
                    "type": "boolean"
                },
                "singleClick": {
                    "description": "open directory on single click, also enables middle click to open in new tab",
                    "type": "boolean"
                },
                "stickySidebar": {
                    "description": "keep sidebar open when navigating",
                    "type": "boolean"
                },
                "themeColor": {
                    "description": "theme color to use: eg. #ff0000, or var(--red), var(--purple), etc",
                    "type": "string"
                },
                "viewMode": {
                    "description": "view mode to use: eg. normal, list, grid, or compact",
                    "type": "string"
                }
            }
        },
        "settings.UserDefaultsPermissions": {
            "type": "object",
            "properties": {
                "admin": {
                    "description": "allow admin access",
                    "type": "boolean"
                },
                "api": {
                    "description": "allow api access",
                    "type": "boolean"
                },
                "create": {
                    "description": "allow creating or uploading files",
                    "type": "boolean"
                },
                "delete": {
                    "description": "allow deleting files",
                    "type": "boolean"
                },
                "download": {
                    "description": "allow downloading files",
                    "type": "boolean"
                },
                "modify": {
                    "description": "allow modifying files",
                    "type": "boolean"
                },
                "realtime": {
                    "description": "allow realtime updates",
                    "type": "boolean"
                },
                "share": {
                    "description": "allow sharing files",
                    "type": "boolean"
                }
            }
        },
        "settings.UserDefaultsPreview": {
            "type": "object",
            "properties": {
                "audio": {
                    "description": "show thumbnails for audio files",
                    "type": "boolean"
                },
                "autoplayMedia": {
                    "description": "autoplay media files in preview",
                    "type": "boolean"
                },
                "defaultMediaPlayer": {
                    "description": "disable the styled feature-rich media player for browser default",
                    "type": "boolean"
                },
                "disableHideSidebar": {
                    "description": "keep sidebar open when previewing files",
                    "type": "boolean"
                },
                "folder": {
                    "description": "show thumbnails for folders that have previewable contents",
                    "type": "boolean"
                },
                "highQuality": {
                    "description": "deprecated: always true in v1.3.0+",
                    "type": "boolean"
                },
                "image": {
                    "description": "show thumbnails for image files",
                    "type": "boolean"
                },
                "models": {
                    "description": "show live thumbnails for 3D models files",
                    "type": "boolean"
                },
                "motionVideoPreview": {
                    "description": "show multiple frames for videos in thumbnail preview when hovering",
                    "type": "boolean"
                },
                "office": {
                    "description": "show thumbnails for office files",
                    "type": "boolean"
                },
                "popup": {
                    "description": "show larger popup preview when hovering over thumbnail",
                    "type": "boolean"
                },
                "video": {
                    "description": "show thumbnails for video files",
                    "type": "boolean"
                }
            }
        },
...
        "users.FileLoading": {
            "type": "object",
            "properties": {
                "clearAll": {
                    "type": "boolean"
                },
                "downloadChunkSizeMb": {
                    "type": "integer"
                },
                "maxConcurrentUpload": {
                    "type": "integer"
                },
                "uploadChunkSizeMb": {
                    "type": "integer"
                }
            }
        },
...
    }
}